+++ lib/binary_types.dart
@@ -1,6 +1,7 @@
 library binary_types;
 
 import 'dart:collection';
+import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
 
@@ -21,7 +22,7 @@ part 'src/errors.dart';
 part 'src/floating_points.dart';
 part 'src/function.dart';
 part 'src/integers.dart';
-part 'src/kinds.dart';
+part 'src/kind.dart';
 part 'src/npe.dart';
 part 'src/object.dart';
 part 'src/physical_data.dart';