+++ lib/src/kind.dart
@@ -1,0 +1,21 @@
+part of binary_types;
+
+enum BinaryKind {
+  ARRAY,
+  BOOL,
+  DOUBLE,
+  ENUM,
+  FLOAT,
+  FUNCTION,
+  POINTER,
+  SINT16,
+  SINT32,
+  SINT64,
+  SINT8,
+  STRUCT,
+  UINT16,
+  UINT32,
+  UINT64,
+  UINT8,
+  VOID
+}